chore: move providers testing
